+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/VariantGroupProcessor.php
@@ -8,7 +8,9 @@ use Akeneo\Bundle\StorageUtilsBundle\Repository\IdentifiableObjectRepositoryInte
 use Doctrine\Common\Collections\ArrayCollection;
 use Pim\Bundle\CatalogBundle\Manager\ProductTemplateMediaManager;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
+use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
+use Pim\Bundle\TransformBundle\Builder\FieldNameBuilder;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Validator\ValidatorInterface;
@@ -50,6 +52,9 @@ class VariantGroupProcessor extends AbstractProcessor
     /** @var string */
     protected $format;
 
+    /** @var FieldNameBuilder */
+    protected $fieldNameBuilder;
+
     /**
      * @param IdentifiableObjectRepositoryInterface $groupRepository
      * @param DenormalizerInterface                 $denormalizer
@@ -57,6 +62,7 @@ class VariantGroupProcessor extends AbstractProcessor
      * @param ObjectDetacherInterface               $detacher
      * @param NormalizerInterface                   $normalizer
      * @param ProductTemplateMediaManager           $templateMediaManager
+     * @param FieldNameBuilder                      $fieldNameBuilder
      * @param string                                $groupClass
      * @param string                                $templateClass
      * @param string                                $format
@@ -68,6 +74,7 @@ class VariantGroupProcessor extends AbstractProcessor
         ObjectDetacherInterface $detacher,
         NormalizerInterface $normalizer,
         ProductTemplateMediaManager $templateMediaManager,
+        FieldNameBuilder $fieldNameBuilder,
         $groupClass,
         $templateClass,
         $format
@@ -75,6 +82,7 @@ class VariantGroupProcessor extends AbstractProcessor
         parent::__construct($groupRepository, $denormalizer, $validator, $detacher, $groupClass);
         $this->normalizer           = $normalizer;
         $this->templateMediaManager = $templateMediaManager;
+        $this->fieldNameBuilder     = $fieldNameBuilder;
         $this->templateClass        = $templateClass;
         $this->format               = $format;
     }
@@ -145,7 +153,7 @@ class VariantGroupProcessor extends AbstractProcessor
     {
         $valuesData = $this->filterVariantGroupData($groupData, false);
         if (!empty($valuesData)) {
-            $values = $this->denormalizeValuesFromItemData($valuesData);
+            $values = $this->denormalizeValuesFromItemData($valuesData, $variantGroup->getProductTemplate());
             $this->validateValues($variantGroup, $values, $groupData);
             $template = $this->getProductTemplate($variantGroup);
             $template->setValues($values);
@@ -212,22 +220,26 @@ class VariantGroupProcessor extends AbstractProcessor
     }
 
     /**
-     * Filter empty values then denormalize the product values objects from CSV fields
+     * Filter empty values that are not used in a template then denormalize the product values objects from CSV fields
      *
-     * @param array $rawProductValues
+     * @param array                    $rawProductValues
+     * @param ProductTemplateInterface $template
      *
      * @return ProductValueInterface[]
      */
-    protected function denormalizeValuesFromItemData(array $rawProductValues)
+    protected function denormalizeValuesFromItemData(array $rawProductValues, ProductTemplateInterface $template = null)
     {
-        $nonEmptyValues = $rawProductValues;
-        foreach ($nonEmptyValues as $index => $data) {
-            if (trim($data) === "") {
-                unset($nonEmptyValues[$index]);
+        $templateCodes = null !== $template ? array_keys($template->getValuesData()) : [];
+
+        foreach ($rawProductValues as $index => $data) {
+            $attributeInfos = $this->fieldNameBuilder->extractAttributeFieldNameInfos($index);
+            $attribute = $attributeInfos['attribute'];
+            if ("" === trim($data) && !in_array($attribute->getCode(), $templateCodes)) {
+                unset($rawProductValues[$index]);
             }
         }
 
-        return $this->denormalizer->denormalize($nonEmptyValues, 'ProductValue[]', 'csv');
+        return $this->denormalizer->denormalize($rawProductValues, 'ProductValue[]', 'csv');
     }
 
     /**